add testing for ensure_relative_file_path!
