+++ src/python/pants/engine/BUILD
@@ -55,8 +55,8 @@ python_library(
     ':selectors',
     'src/python/pants/base:project_tree',
     'src/python/pants/option',
+    'src/python/pants/util:dirutil',
     'src/python/pants/util:meta',
-    'src/python/pants/util:strutil',
   ],
   tags = {'partially_type_checked'},
 )
